pyproject.toml
[project]
name = "immunis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "google-genai>=1.53.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "typer>=0.20.0",
    "uvicorn>=0.38.0",
]
[project.scripts]
immunis-cli = "immunis.cli:main"
